requirements.txt
pygame
websockets
python-chess